+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 A more detailed view of the changes can be found in the {commit messages}[https://github.com/salesking/sk_php_sdk/commits/]
 
+2012-09
+* add HTTP Basic Auth for login with user-email and password
 
 2012-04
 * initial version